refactor upload button to fix ref conflict
